Removing old context file and cleaning up.
