+++ pages/calendar/[...date].tsx
@@ -4,7 +4,6 @@ import { useRouter } from "next/router";
 import ErrorPage from "next/error";
 import Calender from "../../components/calender";
 import { CalenderContextProvider } from "../../contexts/CalenderContext";
-import { NewCalenderContextProvider } from "../../contexts/NewCalenderContext";
 
 interface UpdateCalendarProps {
   year: number;
@@ -72,9 +71,7 @@ const DateRoute: React.FC<unknown> = () => {
   return (
     <Box textAlign="center" w="100%" h="auto" pt="50px" pb="10vh">
       <CalenderContextProvider>
-        <NewCalenderContextProvider>
-          <Calender {...date} />
-        </NewCalenderContextProvider>
+        <Calender {...date} />
       </CalenderContextProvider>
     </Box>
   );
